django/pyproject.toml
[tool.poetry]
name = "spatiallab"
version = "0.1.0"
description = "A web based GIS application"
authors = ["Zach Flanders <zachflanders@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11" 
django = "^5.0" 
gunicorn = "^20.1.0"  
psycopg2-binary = "^2.9.6"
django-environ = "^0.11.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.4.0"
google-cloud-storage = "^2.18.2"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
pyproj = "^3.6.1"
shapely = "^2.0.6"
fiona = "^1.9.6"
google-auth = "^2.34.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
djangorestframework-simplejwt = "^5.3.1"
django-redis = "^5.4.0"
django-ordered-model = "^3.7.4"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0" 
black = "^23.3.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
